+++ packages/backend-common/config.d.ts
@@ -96,6 +96,12 @@ export interface Config {
        * @default database
        */
       pluginDivisionMode?: 'database' | 'schema';
+      /**
+       * Arbitrary config object to pass to knex when initializing
+       * (https://knexjs.org/#Installation-client). Most notable is the debug
+       * and asyncStackTraces booleans
+       */
+      knexConfig?: object;
       /** Plugin specific database configuration and client override */
       plugin?: {
         [pluginId: string]: {
@@ -111,6 +117,14 @@ export interface Config {
            * Defaults to base config if unspecified.
            */
           ensureExists?: boolean;
+          /**
+           * Arbitrary config object to pass to knex when initializing
+           * (https://knexjs.org/#Installation-client). Most notable is the
+           * debug and asyncStackTraces booleans.
+           *
+           * This is merged recursively into the base knexConfig
+           */
+          knexConfig?: object;
         };
       };
     };