+++ packages/core-components/src/layout/Sidebar/Bar.tsx
@@ -17,10 +17,12 @@
 import { makeStyles } from '@material-ui/core/styles';
 import useMediaQuery from '@material-ui/core/useMediaQuery';
 import clsx from 'clsx';
-import React, { useRef, useState, useContext, PropsWithChildren } from 'react';
+import React, { useState, useContext, PropsWithChildren, useRef } from 'react';
 import { sidebarConfig, SidebarContext } from './config';
 import { BackstageTheme } from '@backstage/theme';
 import { SidebarPinStateContext } from './Page';
+import DoubleArrowRight from './icons/DoubleArrowRight';
+import DoubleArrowLeft from './icons/DoubleArrowLeft';
 
 export type SidebarClassKey = 'root' | 'drawer' | 'drawerOpen';
 
@@ -46,7 +48,6 @@ const useStyles = makeStyles<BackstageTheme>(
       msOverflowStyle: 'none',
       scrollbarWidth: 'none',
       width: sidebarConfig.drawerWidthClosed,
-      borderRight: `1px solid #383838`,
       transition: theme.transitions.create('width', {
         easing: theme.transitions.easing.sharp,
         duration: theme.transitions.duration.shortest,
@@ -57,6 +58,19 @@ const useStyles = makeStyles<BackstageTheme>(
       '&::-webkit-scrollbar': {
         display: 'none',
       },
+    },
+    expandButton: {
+      background: 'none',
+      border: 'none',
+      color: theme.palette.navigation.color,
+      width: '100%',
+      cursor: 'pointer',
+      position: 'relative',
+      height: 48,
+    },
+    arrows: {
+      position: 'absolute',
+      right: 10,
     },
     drawerOpen: {
       width: sidebarConfig.drawerWidthOpen,
@@ -78,10 +92,12 @@ enum State {
 type Props = {
   openDelayMs?: number;
   closeDelayMs?: number;
+  disableExpandOnHover?: boolean;
 };
 
 export function Sidebar(props: PropsWithChildren<Props>) {
   const {
+    disableExpandOnHover = false,
     openDelayMs = sidebarConfig.defaultOpenDelayMs,
     closeDelayMs = sidebarConfig.defaultCloseDelayMs,
     children,
@@ -132,18 +148,27 @@ export function Sidebar(props: PropsWithChildren<Props>) {
 
   const isOpen = (state === State.Open && !isSmallScreen) || isPinned;
 
+  const setOpen = (open: boolean) => {
+    if (open) {
+      handleOpen();
+    } else {
+      handleClose();
+    }
+  };
+
   return (
     <div
       className={classes.root}
-      onMouseEnter={handleOpen}
-      onFocus={handleOpen}
-      onMouseLeave={handleClose}
-      onBlur={handleClose}
       data-testid="sidebar-root"
+      onMouseEnter={disableExpandOnHover ? () => {} : handleOpen}
+      onFocus={disableExpandOnHover ? () => {} : handleOpen}
+      onMouseLeave={disableExpandOnHover ? () => {} : handleClose}
+      onBlur={disableExpandOnHover ? () => {} : handleClose}
     >
       <SidebarContext.Provider
         value={{
           isOpen,
+          setOpen,
         }}
       >
         <div
@@ -157,3 +182,39 @@ export function Sidebar(props: PropsWithChildren<Props>) {
     </div>
   );
 }
+
+/**
+ * A button which allows you to expand the sidebar when clicked.
+ * Use optionally to replace sidebar's expand-on-hover feature with expand-on-click.
+ *
+ * @public
+ */
+export const SidebarExpandButton = () => {
+  const classes = useStyles();
+  const { isOpen, setOpen } = useContext(SidebarContext);
+  const { isPinned } = useContext(SidebarPinStateContext);
+  const isSmallScreen = useMediaQuery<BackstageTheme>(theme =>
+    theme.breakpoints.down('md'),
+  );
+
+  const handleClick = () => {
+    setOpen(!isOpen);
+  };
+
+  if (isPinned || isSmallScreen) {
+    return null;
+  }
+
+  return (
+    <button
+      onClick={handleClick}
+      className={classes.expandButton}
+      aria-label="Expand Sidebar"
+      data-testid="sidebar-expand-button"
+    >
+      <div className={classes.arrows}>
+        {isOpen ? <DoubleArrowLeft /> : <DoubleArrowRight />}
+      </div>
+    </button>
+  );
+};