+++ packages/theme/src/themes.ts
@@ -76,6 +76,12 @@ export const lightTheme = createTheme({
       indicator: '#9BF0E1',
       color: '#b5b5b5',
       selectedColor: '#FFF',
+      navItem: {
+        hoverBackground: '#404040',
+      },
+      submenu: {
+        background: '#404040',
+      },
     },
     pinSidebarButton: {
       icon: '#181818',
@@ -151,6 +157,12 @@ export const darkTheme = createTheme({
       indicator: '#9BF0E1',
       color: '#b5b5b5',
       selectedColor: '#FFF',
+      navItem: {
+        hoverBackground: '#404040',
+      },
+      submenu: {
+        background: '#404040',
+      },
     },
     pinSidebarButton: {
       icon: '#404040',