+++ plugins/kubernetes/src/components/Pods/PodsTable.test.tsx
@@ -18,36 +18,19 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import * as pod from './__fixtures__/pod.json';
 import * as crashingPod from './__fixtures__/crashing-pod.json';
-import { TableColumn } from '@backstage/core-components';
 import { wrapInTestApp } from '@backstage/test-utils';
-import { V1Pod } from '@kubernetes/client-node';
-import { PodsTable } from './PodsTable';
-import { containersReady, totalRestarts } from '../../utils/pod';
-
-const extraColumns: TableColumn<V1Pod>[] = [
-  {
-    title: 'containers ready',
-    align: 'center',
-    render: containersReady,
-  },
-  {
-    title: 'total restarts',
-    align: 'center',
-    render: totalRestarts,
-    type: 'numeric',
-  },
-];
+import { PodsTable, READY_COLUMNS, RESOURCE_COLUMNS } from './PodsTable';
+import { kubernetesProviders } from '../../hooks/test-utils';
+import { ClientPodStatus } from '@backstage/plugin-kubernetes-common';
 
 const customColumns: TableColumn<V1Pod>[] = [
   {
     title: 'custom_column',
     align: 'center',
-    render: containersReady,
   },
   {
     title: 'custom_column_2',
     align: 'center',
-    render: totalRestarts,
   },
 ];
 
@@ -71,7 +54,7 @@ describe('PodsTable', () => {
   it('should render pod with extra columns', async () => {
     const { getByText } = render(
       wrapInTestApp(
-        <PodsTable pods={[pod as any]} extraColumns={extraColumns} />,
+        <PodsTable pods={[pod as any]} extraColumns={[READY_COLUMNS]} />,
       ),
     );
 
@@ -89,11 +72,101 @@ describe('PodsTable', () => {
     expect(getByText('0')).toBeInTheDocument();
     expect(getByText('OK')).toBeInTheDocument();
   });
+  it('should render pod, with metrics context', async () => {
+    const podNameToClientPodStatus = new Map<string, ClientPodStatus>();
 
+    podNameToClientPodStatus.set('dice-roller-6c8646bfd-2m5hv', {
+      memory: {
+        currentUsage: '1069056',
+        requestTotal: '67108864',
+        limitTotal: '134217728',
+      },
+      cpu: {
+        currentUsage: 0.4966115,
+        requestTotal: 0.05,
+        limitTotal: 0.05,
+      },
+    } as any);
+
+    const wrapper = kubernetesProviders(
+      undefined,
+      undefined,
+      podNameToClientPodStatus,
+    );
+    const { getByText } = render(
+      wrapper(
+        wrapInTestApp(
+          <PodsTable
+            pods={[pod as any]}
+            extraColumns={[READY_COLUMNS, RESOURCE_COLUMNS]}
+          />,
+        ),
+      ),
+    );
+
+    // titles
+    expect(getByText('name')).toBeInTheDocument();
+    expect(getByText('phase')).toBeInTheDocument();
+    expect(getByText('containers ready')).toBeInTheDocument();
+    expect(getByText('total restarts')).toBeInTheDocument();
+    expect(getByText('status')).toBeInTheDocument();
+    expect(getByText('CPU usage %')).toBeInTheDocument();
+    expect(getByText('Memory usage %')).toBeInTheDocument();
+
+    // values
+    expect(getByText('dice-roller-6c8646bfd-2m5hv')).toBeInTheDocument();
+    expect(getByText('Running')).toBeInTheDocument();
+    expect(getByText('1/1')).toBeInTheDocument();
+    expect(getByText('0')).toBeInTheDocument();
+    expect(getByText('OK')).toBeInTheDocument();
+    expect(getByText('requests: 99%')).toBeInTheDocument();
+    expect(getByText('limits: 99%')).toBeInTheDocument();
+    expect(getByText('requests: 1%')).toBeInTheDocument();
+    expect(getByText('limits: 0%')).toBeInTheDocument();
+  });
+  it('should render placehoplder when empty metrics context', async () => {
+    const podNameToClientPodStatus = new Map<string, ClientPodStatus>();
+
+    const wrapper = kubernetesProviders(
+      undefined,
+      undefined,
+      podNameToClientPodStatus,
+    );
+    const { getByText, getAllByText } = render(
+      wrapper(
+        wrapInTestApp(
+          <PodsTable
+            pods={[pod as any]}
+            extraColumns={[READY_COLUMNS, RESOURCE_COLUMNS]}
+          />,
+        ),
+      ),
+    );
+
+    // titles
+    expect(getByText('name')).toBeInTheDocument();
+    expect(getByText('phase')).toBeInTheDocument();
+    expect(getByText('containers ready')).toBeInTheDocument();
+    expect(getByText('total restarts')).toBeInTheDocument();
+    expect(getByText('status')).toBeInTheDocument();
+    expect(getByText('CPU usage %')).toBeInTheDocument();
+    expect(getByText('Memory usage %')).toBeInTheDocument();
+
+    // values
+    expect(getByText('dice-roller-6c8646bfd-2m5hv')).toBeInTheDocument();
+    expect(getByText('Running')).toBeInTheDocument();
+    expect(getByText('1/1')).toBeInTheDocument();
+    expect(getByText('0')).toBeInTheDocument();
+    expect(getByText('OK')).toBeInTheDocument();
+    expect(getAllByText('unknown')).toHaveLength(2);
+  });
   it('should render crashing pod with extra columns', async () => {
     const { getByText, getAllByText } = render(
       wrapInTestApp(
-        <PodsTable pods={[crashingPod as any]} extraColumns={extraColumns} />,
+        <PodsTable
+          pods={[crashingPod as any]}
+          extraColumns={[READY_COLUMNS]}
+        />,
       ),
     );
 
@@ -116,10 +189,13 @@ describe('PodsTable', () => {
     expect(getAllByText('CrashLoopBackOff')).toHaveLength(2);
   });
   it('should render additional custom columns', async () => {
-    const columnsToAdd = extraColumns.concat(customColumns);
     const { getByText, getAllByText } = render(
       wrapInTestApp(
-        <PodsTable pods={[pod as any]} extraColumns={columnsToAdd} />,
+        <PodsTable
+          pods={[pod as any]}
+          extraColumns={[READY_COLUMNS, RESOURCE_COLUMNS]}
+          customColumns={customColumns}
+        />,
       ),
     );
 
@@ -129,14 +205,17 @@ describe('PodsTable', () => {
     expect(getByText('containers ready')).toBeInTheDocument();
     expect(getByText('total restarts')).toBeInTheDocument();
     expect(getByText('status')).toBeInTheDocument();
+    expect(getByText('CPU usage %')).toBeInTheDocument();
+    expect(getByText('Memory usage %')).toBeInTheDocument();
     expect(getByText('custom_column')).toBeInTheDocument();
     expect(getByText('custom_column_2')).toBeInTheDocument();
 
     // values
     expect(getByText('dice-roller-6c8646bfd-2m5hv')).toBeInTheDocument();
     expect(getByText('Running')).toBeInTheDocument();
-    expect(getAllByText('1/1')).toHaveLength(2);
-    expect(getAllByText('0')).toHaveLength(2);
+    expect(getByText('1/1')).toBeInTheDocument();
+    expect(getByText('0')).toBeInTheDocument();
     expect(getByText('OK')).toBeInTheDocument();
+    expect(getAllByText('unknown')).toHaveLength(2);
   });
 });