+++ plugins/kubernetes/src/types/types.ts
@@ -21,7 +21,7 @@ import {
   V1HorizontalPodAutoscaler,
   V1Service,
   V1ConfigMap,
-  ExtensionsV1beta1Ingress,
+  V1Ingress,
   V1Job,
   V1CronJob,
 } from '@kubernetes/client-node';
@@ -36,7 +36,7 @@ export interface DeploymentResources {
 export interface GroupedResponses extends DeploymentResources {
   services: V1Service[];
   configMaps: V1ConfigMap[];
-  ingresses: ExtensionsV1beta1Ingress[];
+  ingresses: V1Ingress[];
   jobs: V1Job[];
   cronJobs: V1CronJob[];
   customResources: any[];