+++ packages/app/CHANGELOG.md
@@ -1,5 +1,25 @@
 # example-app
 
+## 0.2.55
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/plugin-azure-devops@0.1.5
+  - @backstage/core-components@0.7.6
+  - @backstage/theme@0.2.14
+  - @backstage/plugin-user-settings@0.3.12
+  - @backstage/plugin-api-docs@0.6.16
+  - @backstage/cli@0.10.0
+  - @backstage/plugin-circleci@0.2.30
+  - @backstage/core-plugin-api@0.2.2
+  - @backstage/plugin-search@0.5.0
+  - @backstage/plugin-tech-insights@0.1.0
+  - @backstage/core-app-api@0.1.24
+  - @backstage/plugin-kubernetes@0.4.22
+  - @backstage/plugin-scaffolder@0.11.13
+  - @backstage/plugin-techdocs@0.12.8
+
 ## 0.2.53
 
 ### Patch Changes