+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @backstage/cli
 
+## 0.10.0
+
+### Minor Changes
+
+- ea99ef5198: Remove the `backend:build-image` command from the CLI and added more deprecation warnings to other deprecated fields like `--lax` and `remove-plugin`
+
+### Patch Changes
+
+- e7230ef814: Bump react-dev-utils to v12
+- 416b68675d: build(dependencies): bump `style-loader` from 1.2.1 to 3.3.1
+- Updated dependencies
+  - @backstage/config-loader@0.8.1
+
 ## 0.9.1
 
 ### Patch Changes