+++ packages/integration/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @backstage/integration
 
+## 0.6.10
+
+### Patch Changes
+
+- 47619da24c: Narrow the types returned by the request option functions, to only the specifics that they actually do return. The reason for this change is that a full `RequestInit` is unfortunate to return because it's different between `cross-fetch` and `node-fetch`.
+
 ## 0.6.9
 
 ### Patch Changes